# [Codegen][ROCDL] Use buffer fat pointers where possible