--- a/compiler/src/iree/compiler/Codegen/Common/CMakeLists.txt
+++ b/compiler/src/iree/compiler/Codegen/Common/CMakeLists.txt
@@ -164,6 +164,7 @@ iree_cc_library(
     IREELinalgTransformDialect
     LLVMSupport
     MLIRAMDGPUDialect
+    MLIRAMDGPUTransforms
     MLIRAMDGPUUtils
     MLIRAffineAnalysis
     MLIRAffineDialect