--- a/compiler/src/iree/compiler/Codegen/Common/VerifyWorkgroupDistribution.cpp
+++ b/compiler/src/iree/compiler/Codegen/Common/VerifyWorkgroupDistribution.cpp
@@ -8,6 +8,7 @@
 #include "iree/compiler/Codegen/Dialect/Codegen/IR/IREECodegenAttrs.h"
 #include "iree/compiler/Codegen/Utils/GPUUtils.h"
 #include "iree/compiler/Dialect/HAL/IR/HALTypes.h"
+#include "mlir/Dialect/AMDGPU/IR/AMDGPUDialect.h"
 #include "mlir/IR/Visitors.h"
 #include "mlir/Interfaces/FunctionInterfaces.h"
 #include "mlir/Interfaces/SideEffectInterfaces.h"
@@ -41,9 +42,6 @@ struct VerifyWorkgroupDistributionPass final
       return;
     }
 
-    auto globalAddressSpace = IREE::HAL::DescriptorTypeAttr::get(
-        &getContext(), IREE::HAL::DescriptorType::StorageBuffer);
-
     // Walk in PreOrder so that parent operations are visited before children,
     // thus allowing all operations contained within workgroup foralls to be
     // skipped.
@@ -63,8 +61,10 @@ struct VerifyWorkgroupDistributionPass final
             continue;
           }
 
-          // Writes to non-global memory are fine.
-          if (type.getMemorySpace() != globalAddressSpace) {
+          // Writes to non-global memory are fine. hasGlobalMemoryAddressSpace()
+          // treats the unspecified address space as global, but we don't have
+          // that assumption yet.
+          if (!type.getMemorySpace() || !hasGlobalMemoryAddressSpace(type)) {
             continue;
           }