--- a/compiler/src/iree/compiler/Codegen/Dialect/GPU/IR/IREEGPUDialect.td
+++ b/compiler/src/iree/compiler/Codegen/Dialect/GPU/IR/IREEGPUDialect.td
@@ -30,6 +30,9 @@ def IREEGPU_Dialect : Dialect {
   }];
   let useDefaultAttributePrinterParser = 1;
 
+  let discardableAttrs = (ins
+    "::mlir::UnitAttr":$use_rocdl_buffer_instructions
+  );
   let extraClassDeclaration = [{
     void registerAttributes();
   }];