--- a/compiler/src/iree/compiler/Codegen/LLVMGPU/BUILD.bazel
+++ b/compiler/src/iree/compiler/Codegen/LLVMGPU/BUILD.bazel
@@ -108,6 +108,7 @@ iree_compiler_cc_library(
         "LLVMGPUVectorToGPU.cpp",
         "Passes.cpp",
         "ROCDLAnnotateKernelForTranslation.cpp",
+        "ROCDLConfigureBufferInstructions.cpp",
         "ROCDLKernelConfig.cpp",
         "ROCDLLowerExecutableTarget.cpp",
         "ROCDLSelectLoweringStrategy.cpp",