--- a/compiler/src/iree/compiler/Codegen/LLVMGPU/ROCDLPasses.td
+++ b/compiler/src/iree/compiler/Codegen/LLVMGPU/ROCDLPasses.td
@@ -19,6 +19,26 @@ def ROCDLAnnotateKernelForTranslationPass : Pass<
   let dependentDialects = ["ROCDL::ROCDLDialect"];
 }
 
+def ROCDLConfigureBufferInstructionsPass :
+    InterfacePass<"iree-rocdl-configure-buffer-instructions", "mlir::FunctionOpInterface"> {
+  let summary = "Determine which subspans can be implemneted with buffer fat pointers";
+  let description = [{
+    Analyze hal.interface.binding.subspan operations within a dispatch to determine
+    whether
+    - The size of the binding itself, taken as a region of memory, is under 2 GB
+     (the limit for indexing a buffer descriptor is uint32_t max, but a lot of
+     intermediate code assumes indices are signed, se this limit is conservative) and
+    - That the offset argument to the subspan is workgroup-uniform, since it will
+     be folded into the base offset of the descriptor.
+
+    Note that this pass only adds annotations to subspans - the descriptor will be
+    constructed during bufferization based on annotations from this pass.
+
+    This pass is a no-op on non-ROCDL targets.
+  }];
+  let dependentDialects = ["::mlir::iree_compiler::IREE::GPU::IREEGPUDialect"];
+}
+
 def ROCDLLowerExecutableTargetPass : InterfacePass<
     "iree-rocdl-lower-executable-target", "mlir::FunctionOpInterface"> {
   let summary = "Lower an IREE hal.executable.variant op using a suitable "