This commit enables using the `#amdgpu.address_space<fat_raw_buffer>` address space for memref subspans in our kernels in cases where it's safe to do so. This both gains us immediate performance improvements (-2ms on unet) and enables further performance improvements on MI-300 through the use of the cache swizzling mechanism.

Note: this PR currently depends on a bunch of open upstream work.

# Context

AMD GPUs, on top of having normal `global_{load,store,...}_T` instructions, which take a 64-bit pointer, also have `buffer_*` instructions, which use the texture unit to enable more complex logic.

These instructions, for our purposes, take a "buffer resource" (a base pointer into global memory + flags) and an offset. The resource takes 4 **scalar** registers (and thus must be wave-uniform) and the offset is 1 **vector** register, which is added to the base pointer to form the effective address of the buffer access.

This means that we can use these buffer operations if we can be assured that the non-uniform part of our accesses never exceeds 2 GB from the start of whatever we decide the base pointer is.

IREE will often pass in N copies of the same pointer and then take global offsets from that pointer to locate the logical bindings/tensors within the model's data. Fortunately, these offsets are globally uniform.

Therefore, we can conservatively construct buffer descriptors by finding the start of a given bound tensor/memref (which is always in a contiguous layout) and using that as a base pointer. Then, if we know that the memref is 2 GB or smaller, we can use the buffer instructions to access it.

# How this patch works

1. Before bufferization, we run iree-rocdl-configure-buffer-instructions / ROCDLConfigureBufferInstructions , which will annotate `hal.interface.binding.subspan` operations with the key ` iree_gpu.use_rocdl_buffer_instructions` as a discardable attribute on the subspans *if* we can prove that said subspans are safe to use as buffer instructions
2. During bufferization, if ` iree_gpu.use_rocdl_buffer_instructions` is set, then we use the `#amdgpu.address_space<buffer_fat_pointer>` memory space instead of the `#hal.descriptor_type<*>` one. This required refactoring a few passes that were using `#hal.descriptor_type` as a synonym for "global memory" in tests, but that was a fairly small change
3. Finally, in `iree-codegen-convert-hal-descriptor-type-to-gpu-address-space`, we implement the buffer fat pointer ABI by converting these buffer bindings to normal global bindings and then adding an `amdgpu.fat_raw_buffer_cast` with the `resetOffset` option set to fold in the global byte offset. (I'm not entirely happy with this part and maybe we should move it to bufferization)

So it's recorded: +~3% on SDXL on MI-300 from this