@krzysz00 2025-02-05
Ok, so I ran the UNet kernel gemm benchmarks from iree-kernel-benchmark on sharkmi300x-3. I'm not sure I believe these numbers, given how wide the variance is, 3% +- 8% is ... not bad, I guess?

I'll take a look at some assembly diffs, see if anything jumps out at me here.

Stat | Value
-- | --
Average | 3.036367
Std. Dev. | 8.519886
Min | -21.7395
Max | 38.63633
25%ile | -1.08869
Median | 1.951279
75%ile | 4.742739

<details>
<summary> Raw TFlops Numbers</summary>

Problem | Changed Tflops | Main Tflops | % Inc
-- | -- | -- | --
gemm_2048_10240_1280_f16_f32 | 298.2616 | 296.6138 | 0.555537
gemm_2048_1280_1280_f16_f32 | 94.5195 | 87.1544 | 8.450635
gemm_2048_1280_5120_f16_f32 | 168.8273 | 166.7301 | 1.257841
gemm_128_1280_2048_f16_f32 | 9.7259 | 12.4276 | -21.7395
gemm_8192_640_640_f16_f32 | 100.1625 | 93.2068 | 7.462653
gemm_8192_640_2560_f16_f32 | 194.5184 | 197.379 | -1.44929
gemm_8192_5120_640_f16_f32 | 299.9279 | 298.2616 | 0.558671
gemm_64_640_2048_f16_f32 | 2.6214 | 2.6631 | -1.56584
gemm_64_1280_2048_f16_f32 | 5.412 | 5.084 | 6.451613
gemm_1024_1280_1280_f16_f32 | 51.6222 | 51.6222 | 0
gemm_1024_1280_5120_f16_f32 | 104.0448 | 110.9237 | -6.20147
gemm_1024_10240_1280_f16_f32 | 225.576 | 223.6962 | 0.840336
gemm_4096_640_640_f16_f32 | 54.1201 | 54.1201 | 0
gemm_4096_640_2560_f16_f32 | 130.3085 | 126.6205 | 2.912641
gemm_4096_5120_640_f16_f32 | 227.4877 | 221.8475 | 2.542377
gemm_2048_10240_1280_f16_f32_tB | 367.7198 | 353.2045 | 4.109602
gemm_2048_1280_1280_f16_f32_tB | 95.8698 | 94.5195 | 1.428594
gemm_2048_1280_5120_f16_f32_tB | 175.448 | 175.448 | 0
gemm_128_1280_2048_f16_f32_tB | 12.4276 | 11.9837 | 3.704198
gemm_8192_640_640_f16_f32_tB | 98.6895 | 110.0145 | -10.2941
gemm_8192_640_2560_f16_f32_tB | 204.9126 | 211.3665 | -3.05342
gemm_8192_5120_640_f16_f32_tB | 365.2183 | 350.896 | 4.081637
gemm_64_640_2048_f16_f32_tB | 2.8436 | 2.8926 | -1.69398
gemm_64_1280_2048_f16_f32_tB | 6.2138 | 5.5924 | 11.11151
gemm_1024_1280_1280_f16_f32_tB | 54.1201 | 59.9186 | -9.6773
gemm_1024_1280_5120_f16_f32_tB | 138.3688 | 122.0161 | 13.40208
gemm_1024_10240_1280_f16_f32_tB | 298.2616 | 250.8743 | 18.88886
gemm_4096_640_640_f16_f32_tB | 76.2601 | 55.0073 | 38.63633
gemm_4096_640_2560_f16_f32_tB | 150.8064 | 134.2177 | 12.35955
gemm_4096_5120_640_f16_f32_tB | 305.0403 | 250.8743 | 21.59089
gemm_2048_10240_1280_bf16_f32 | 305.0403 | 293.3721 | 3.97727
gemm_2048_1280_1280_bf16_f32 | 94.5195 | 84.9479 | 11.26761
gemm_2048_1280_5120_bf16_f32 | 168.8273 | 161.7081 | 4.402501
gemm_128_1280_2048_bf16_f32 | 11.7735 | 12.662 | -7.01706
gemm_8192_640_640_bf16_f32 | 91.93 | 94.5195 | -2.73965
gemm_8192_640_2560_bf16_f32 | 197.379 | 190.3798 | 3.67644
gemm_8192_5120_640_bf16_f32 | 299.9279 | 290.2005 | 3.351958
gemm_64_640_2048_bf16_f32 | 2.6214 | 2.5041 | 4.684318
gemm_64_1280_2048_bf16_f32 | 5.7852 | 4.9345 | 17.23984
gemm_1024_1280_1280_bf16_f32 | 51.6222 | 50.0812 | 3.077003
gemm_1024_1280_5120_bf16_f32 | 110.0145 | 104.8576 | 4.918003
gemm_1024_10240_1280_bf16_f32 | 225.576 | 220.0291 | 2.520985
gemm_4096_640_640_bf16_f32 | 54.1201 | 56.8719 | -4.83859
gemm_4096_640_2560_bf16_f32 | 130.3085 | 127.8264 | 1.941774
gemm_4096_5120_640_bf16_f32 | 225.576 | 225.576 | 0
gemm_2048_10240_1280_bf16_f32_tB | 355.5437 | 360.316 | -1.32448
gemm_2048_1280_1280_bf16_f32_tB | 100.1625 | 98.6895 | 1.49256
gemm_2048_1280_5120_bf16_f32_tB | 189.0391 | 176.6023 | 7.042264
gemm_128_1280_2048_bf16_f32_tB | 11.1848 | 11.5705 | -3.33348
gemm_8192_640_640_bf16_f32_tB | 111.8481 | 106.522 | 5
gemm_8192_640_2560_bf16_f32_tB | 206.4888 | 209.7152 | -1.53847
gemm_8192_5120_640_bf16_f32_tB | 350.896 | 344.148 | 1.960784
gemm_64_640_2048_bf16_f32_tB | 2.6631 | 2.7962 | -4.76003
gemm_64_1280_2048_bf16_f32_tB | 5.7852 | 5.6872 | 1.723168
gemm_1024_1280_1280_bf16_f32_tB | 53.261 | 51.6222 | 3.174603
gemm_1024_1280_5120_bf16_f32_tB | 123.1355 | 118.7768 | 3.669656
gemm_1024_10240_1280_bf16_f32_tB | 255.6528 | 253.241 | 0.952373
gemm_4096_640_640_bf16_f32_tB | 67.1089 | 56.8719 | 18.0001
gemm_4096_640_2560_bf16_f32_tB | 135.5735 | 136.9569 | -1.0101
gemm_4096_5120_640_bf16_f32_tB | 253.241 | 253.241 | 0
</details>