@krzysz00 2025-02-05
Peeking at the anonymized assembly diff on `gemm_64_640_2048_f16_f32`, I'm mainly seeing
- The changes from `global_*` to `buffer_*` I expected
- Less or different math as there's a touch less 64-bit arithmetic running raound (though thit is the sort of thing I'd expect to also clean up with aggressive use of `nsw`, `nusw` and other compiler hints, even without buffers)
- Somewhat different instruction schedules (the f16=>f32 conversions moved up a bit, some waitcnts were collapsed into a wait-to-0)

(That's one of the ones that performed mildly worse, I realized afterwards)

Similarly, the diffs for `gemm_1024_10240_1280_f16_f32_tB` - one of the big improvements - is showing ... mainly arithmetic changes, the instructions getting swapped out, and some knock-on scheduling effects.

In conclusion:
a) Huh?
b) We should keep going

... On second thought, my rocMLIR "going to global loads sometimes helps" numbers were from an index=i32 where save universe, so maybe part of we're seeing is just the effects of needing less registers