@kuhar 2025-02-05
@krzysz00 how to interpret those values in? Are negative values an improvement or a regression? What's the unit? How repeatable/noisy is each configuration?