@krzysz00 2025-02-05
@kuhar Right, I left off detail. The unit's % increase in TFlops, since I went with `(tflops_my_change - tflops_main) / tflops_main * 100`. So positive's better.

I don't know how repeatable these are, I'd assumed that the iree-kernel-benchmark scripts do something reasonable to smooth out the worst of the noise