@kuhar 2025-02-05
Ah, makes sense. Looks promising!

> I'd assumed that the iree-kernel-benchmark scripts do something reasonable to smooth out the worst of the noise

It does but we don't report what the noise levels are.

Could you see if this helps with SDXL (unet)? The reported number should be much more stable and meaningful than these micro benchmarks.